% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr.time2.r
\name{aggr.time2}
\alias{aggr.time2}
\title{Aggregate data over time.
Much faster than version 1, but that one is kept for backwards compatibility.
This will aggregate *all* fields in the dataframe, except the POSIX. NA values will be omitted.}
\usage{
aggr.time2(dat, by, f = "mean")
}
\arguments{
\item{dat}{(data)    Dataframe containing everything (data,posix)}

\item{by}{(string)  A time interval. Possible values: day, week, month, year}

\item{f}{(string)  A function. Possible values: mean (default), sum, max, min, sd}
}
\value{
A list containing the $data and the coresponding $posix dates (*first* moment of day/week/month/year is given)
}
\description{
Aggregate data over time.
Much faster than version 1, but that one is kept for backwards compatibility.
This will aggregate *all* fields in the dataframe, except the POSIX. NA values will be omitted.
}
\author{
Wim Verbruggen
}
