% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.moving.r
\name{avg.moving}
\alias{avg.moving}
\title{Moving average based on index values}
\usage{
avg.moving(dat, win = 5)
}
\arguments{
\item{dat}{Data}

\item{win}{Window size (index values)}
}
\description{
Moving average based on index values
}
