% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.mae.r
\name{error.mae}
\alias{error.mae}
\title{Mean absolute error}
\usage{
error.mae(obs, sim, na_remove = TRUE)
}
\arguments{
\item{obs}{the 'observed' data (can also be field data for example)}

\item{sim}{the 'simulated' data (can also be satellite data for example)}

\item{na_remove}{whether to ignore NA values}
}
\value{
Mean absolute error
}
\description{
Mean absolute error
}
