% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.rmsd.r
\name{error.rmsd}
\alias{error.rmsd}
\title{Root mean square deviation}
\usage{
error.rmsd(obs, sim, normalize = "mean", na_remove = TRUE)
}
\arguments{
\item{obs}{the 'observed' data (can also be field data for example)}

\item{sim}{the 'simulated' data (can also be satellite data for example)}

\item{normalize}{the normalizing method: none, mean, range, stdev}

\item{na_remove}{whether to ignore NA values}
}
\value{
Root mean square deviation (fractional)
}
\description{
Root mean square deviation
}
