% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.shift.r
\name{index.shift}
\alias{index.shift}
\title{Shift index of variable by given lag.}
\usage{
index.shift(var, lag, rotate = FALSE)
}
\arguments{
\item{var}{The variable of interest}

\item{lag}{The lag, either positive or negative.}

\item{rotate}{Add dropped values back at other side?}
}
\description{
Shift index of variable by given lag.
}
