% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namex.r
\name{namex}
\alias{namex}
\title{Get the names of an object, so the same as names(), with the added possibility
to exclude certain names.}
\usage{
namex(obj, excl = "")
}
\arguments{
\item{obj}{object}

\item{excl}{list of names to be excluded}
}
\value{
List of names in the object, excluding the ones mentioned
}
\description{
Get the names of an object, so the same as names(), with the added possibility
to exclude certain names.
}
\author{
Wim Verbruggen
}
