% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ggmulti.r
\name{plot.ggmulti}
\alias{plot.ggmulti}
\title{Combine multiple ggplots on one plot}
\usage{
\method{plot}{ggmulti}(..., plotlist = NULL, file, cols = 1,
  layout = NULL)
}
\description{
Combine multiple ggplots on one plot
}
