% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.moving.r
\name{sum.moving}
\alias{sum.moving}
\title{Moving sum based on index values}
\usage{
\method{sum}{moving}(dat, win = 5)
}
\arguments{
\item{dat}{Data}

\item{win}{Window size (index values)}
}
\description{
Moving sum based on index values
}
